fastapi-app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
python-dotenv>=1.0.0
gtts>=2.5.4
openai>=2.11.0
SpeechRecognition>=3.14.4
Pillow>=11.3.0
aiofiles>=24.1.0
pydantic>=2.11.9
pydantic-settings>=2.5.2
python-docx>=1.1.0
elevenlabs>=0.2.27
pydub>=0.25.1
httpx>=0.27.0
chromadb
pysqlite3-binary ; sys_platform == 'linux'
